# cmd/api/docs/swagger.yaml
definitions:
  AttendanceResponse:
    description: Attendance record response
    properties:
      check_in_at:
        type: string
      check_in_latitude:
        type: number
      check_in_longitude:
        type: number
      check_out_at:
        type: string
      check_out_latitude:
        type: number
      check_out_longitude:
        type: number
      created_at:
        type: string
      duration_seconds:
        type: integer
      id:
        type: integer
      is_early_leave:
        type: boolean
      is_late:
        type: boolean
      overtime_seconds:
        type: integer
      shift_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  CheckInRequest:
    description: Check-in request with optional geolocation
    properties:
      client_timestamp:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
  CheckOutRequest:
    description: Check-out request with optional geolocation
    properties:
      attendance_id:
        description: 'Optional: specify which session to close'
        type: integer
      client_timestamp:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  LeaveBalanceResponse:
    properties:
      accrued_days:
        type: number
      available_days:
        type: number
      carryover_days:
        type: number
      leave_type:
        type: string
      used_days:
        type: number
      year:
        type: integer
    type: object
  ListAttendanceResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      records:
        items:
          $ref: '#/definitions/AttendanceResponse'
        type: array
      total:
        type: integer
    type: object
  ManualAttendanceRequest:
    properties:
      check_in_at:
        type: string
      check_out_at:
        type: string
      overtime_seconds:
        type: integer
      reason:
        minLength: 10
        type: string
      user_id:
        type: integer
    required:
    - check_in_at
    - check_out_at
    - reason
    - user_id
    type: object
  ReviewTimeOffRequest:
    properties:
      review_note:
        minLength: 5
        type: string
    required:
    - review_note
    type: object
...
  github_com_waqasmani_go-boilerplate_internal_shared_utils.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.ErrorResponse'
      success:
        type: boolean
      version:
        type: string
    type: object
...
info:
  contact: {}
  description: This is an attendance and employee management server.
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /attendance:
    get:
      consumes:
      - application/json
      description: Retrieves attendance records with pagination and filters
      parameters:
      - description: User ID filter
        in: query
        name: user_id
        type: integer
      - description: Start date (RFC3339)
        in: query
        name: from
        type: string
      - description: End date (RFC3339)
        in: query
        name: to
        type: string
      - description: Status filter
        enum:
        - present
        - on_leave
        - absent
        - manual
        in: query
        name: status
        type: string
      - description: Shift ID filter
        in: query
        name: shift_id
        type: integer
      - default: 20
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/ListAttendanceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: List attendance records
      tags:
      - Attendance
  /attendance/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single attendance record by ID
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/AttendanceResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Get attendance by ID
      tags:
      - Attendance
  /attendance/{id}/manual:
    put:
      consumes:
      - application/json
      description: Allows managers to manually create or edit attendance records
      parameters:
      - description: Attendance ID (0 for new)
        in: path
        name: id
        required: true
        type: integer
      - description: Manual attendance data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ManualAttendanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/AttendanceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Manually edit attendance (Admin/Manager only)
      tags:
      - Attendance
  /attendance/check-in:
    post:
      consumes:
      - application/json
      description: Records employee check-in with optional geolocation
      parameters:
      - description: Check-in details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CheckInRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/AttendanceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Check in to work
      tags:
      - Attendance
  /attendance/check-out:
    post:
      consumes:
      - application/json
      description: Records employee check-out and calculates duration
      parameters:
      - description: Check-out details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CheckOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/AttendanceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Check out from work
      tags:
      - Attendance
  /attendance/leave-balance:
    get:
      consumes:
      - application/json
      description: Retrieves current leave balance for authenticated user
      parameters:
      - description: Leave type
        enum:
        - sick
        - vacation
        - personal
        in: query
        name: leave_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/LeaveBalanceResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Get leave balance
      tags:
      - Attendance
  /attendance/time-off:
    post:
      consumes:
      - application/json
      description: Employee submits a time off request
      parameters:
      - description: Time off request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/TimeOffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/TimeOffResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Request time off
      tags:
      - Time Off
  /attendance/time-off/{id}/approve:
    patch:
      consumes:
      - application/json
      description: Manager approves a pending time off request
      parameters:
      - description: Time off request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review note
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ReviewTimeOffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Approve time off (Manager only)
      tags:
      - Time Off
  /attendance/time-off/{id}/reject:
    patch:
      consumes:
      - application/json
      description: Manager rejects a pending time off request
      parameters:
      - description: Time off request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review note
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ReviewTimeOffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Reject time off (Manager only)
      tags:
      - Time Off
...
  /reports/daily-summary:
    get:
      consumes:
      - application/json
      description: Retrieves attendance summary for a specific date
      parameters:
      - description: Date (YYYY-MM-DD format)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/DailySummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Get daily attendance summary
      tags:
      - Reports
  /reports/employee/{id}/export:
    get:
      consumes:
      - application/json
      description: Exports timesheet data for an employee for a specific period
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD format)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD format)
        in: query
        name: end_date
        required: true
        type: string
      responses:
        "200":
          description: Timesheet CSV file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Export employee timesheet
      tags:
      - Reports
...
securityDefinitions:
  Bearer:
    description: Type "Bearer <your_token>" (include the word Bearer and a space)
    in: header
    name: Authorization
    type: apiKey
  CsrfToken:
    description: Enter your CSRF token
    in: header
    name: X-CSRF-TOKEN
    type: apiKey
swagger: "2.0"